refactor: remove underscore to rename package
